assistant: Extract `ConfigurationView` to its own module
